requirements.txt
alembic==1.8.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environ==1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg==3.0.16
psycopg2
psycopg-binary==3.0.16
psycopg2-binary==2.9.3
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0